-add new Exception

+++ src/main/java/com/example/animalchipization/service/AnimalService.java
@@ -1,12 +1,15 @@
 package com.example.animalchipization.service;
 
 import com.example.animalchipization.domain.*;
-import com.example.animalchipization.dto.*;
-import com.example.animalchipization.exception.AlreadyExistException;
+import com.example.animalchipization.dto.AnimalDTO;
+import com.example.animalchipization.dto.SearchAnimalDTO;
+import com.example.animalchipization.dto.VisitedLocationDTO;
 import com.example.animalchipization.exception.InaccessibleEntityException;
 import com.example.animalchipization.exception.NotFoundException;
 import com.example.animalchipization.repository.*;
 import com.example.animalchipization.util.CriteriaManager;
+import com.example.animalchipization.util.SecurityUtil;
+import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,23 +19,20 @@ import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 @Service
+@AllArgsConstructor
 public class AnimalService {
-    @Autowired
-    private ModelMapper modelMapper;
-    @Autowired
-    private AnimalRepository animalRepository;
-    @Autowired
-    private LocationRepository locationRepository;
-    @Autowired
-    private AnimalTypeRepository animalTypeRepository;
-    @Autowired
-    private VisitedLocationsRepository visitedLocationsRepository;
-    @Autowired
-    private CriteriaManager criteriaManager;
-    @Autowired
-    private EntityRepository entityRepository;
+        private final ModelMapper modelMapper;
+        private final AnimalRepository animalRepository;
+        private final LocationRepository locationRepository;
+        private final AnimalTypeRepository animalTypeRepository;
+        private final VisitedLocationsRepository visitedLocationsRepository;
+        private final AccountRepository accountRepository;
+        private final CriteriaManager criteriaManager;
+        private final EntityRepository entityRepository;
+        private final  AnimalTypeService animalTypeService;
 
     public AnimalDTO getById(Long id){
         Animal animal = animalRepository.findById(id)
@@ -52,6 +52,13 @@ public class AnimalService {
         dto.setAnimalTypesIds(types);
 
         return dto;
+    }
+
+    public List<VisitedLocationDTO> allVisitedLocation(Long id){
+        return visitedLocationsRepository.findAllByAnimal(id).stream()
+                .map(entity -> new VisitedLocationDTO(entity.getId(), entity.getVisitTime(), entity.getLocation().getId()))
+                .toList();
+
     }
 
     public VisitedLocationDTO addVisitedLocation(Long animalId, Long locationId){
@@ -77,21 +84,35 @@ public class AnimalService {
 
     public AnimalDTO createAnimal(AnimalDTO dto){
         Animal animal = new Animal();
-        List<AnimalType> types = animalTypeRepository.findAllByIdIn(dto.getAnimalTypesIds());
 
-        if(types.size() != dto.getAnimalTypesIds().size()){
-            throw new NotFoundException();
+        if(dto.getAnimalTypesIds() == null || dto.getAnimalTypesIds().isEmpty()){
+            throw new IllegalStateException();
         }
+        List<AnimalType> types = animalTypeService.findAllByIds(dto.getAnimalTypesIds());
+
 
         modelMapper.map(dto, animal);
+
+        Account account = accountRepository.findById(dto.getChipperId())
+                .orElseThrow(() -> new NotFoundException());
+
+        Location location = locationRepository.findById(dto.getChippingLocationId())
+                        .orElseThrow(() -> new NotFoundException());
+
+        animal.setChippingLocationId(location);
+        animal.setChipper(account);
         animal.setAnimalTypes(types);
 
         animal = animalRepository.save(animal);
         AnimalDTO animalDTO = modelMapper.map(animal, AnimalDTO.class);
+
         List<Long> ids = animal.getAnimalTypes().stream()
                 .map(type -> type.getId())
                 .toList();
+
         animalDTO.setAnimalTypesIds(ids);
+        animalDTO.setChippingLocationId(animal.getChippingLocationId().getId());
+        animalDTO.setChipperId(account.getId());
 
         return animalDTO;
     }
@@ -112,29 +133,38 @@ public class AnimalService {
         return listDTO;
     }
 
-    public AnimalDTO updateAnimal(Long id, AnimalDTO animalDTO){
+    public AnimalDTO updateAnimal(Long id, AnimalDTO dto){
         Animal animal = animalRepository.getById(id);
 
-        animal.setWeight(animalDTO.getWeight());
-        animal.setLength(animalDTO.getLength());
-        animal.setHeight(animalDTO.getHeight());
-        animal.setGender(animalDTO.getGender());
-        animal.setLifeStatus(animalDTO.getLifeStatus());
-        animal.setChippingDateTime(animalDTO.getChippingDateTime());
-        animal.setChipperId(animalDTO.getChipperId());
-        animal.setChippingLocationId(animalDTO.getChippingLocationId());
-        animal.setDeathDateTime(animalDTO.getDeathDateTime());
-//        animal.setAnimalTypes(animalDTO.getAnimalTypes());
-//        animal.setVisitedLocations(animalDTO.getVisitedLocations());
+        if(dto.getAnimalTypesIds() != null && !dto.getAnimalTypesIds().isEmpty()){
+            List<AnimalType> types = animalTypeService.findAllByIds(dto.getAnimalTypesIds());
+            animal.setAnimalTypes(types);
+        }
+
+        animal.setWeight(dto.getWeight());
+        animal.setLength(dto.getLength());
+        animal.setHeight(dto.getHeight());
+        animal.setGender(dto.getGender());
+        animal.setLifeStatus(dto.getLifeStatus());
+        animal.setChippingDateTime(dto.getChippingDateTime());
+       // animal.setChipperId(animalDTO.getChipperId());
+       // animal.setChippingLocationId(animalDTO.getChippingLocationId());
+        animal.setDeathDateTime(dto.getDeathDateTime());
+
+        //        animal.setVisitedLocations(animalDTO.getVisitedLocations());
 
         Animal saveAnimal = animalRepository.save(animal);
-        AnimalDTO dto = modelMapper.map(saveAnimal, AnimalDTO.class);
-        return dto;
+
+        return modelMapper.map(saveAnimal, AnimalDTO.class);
     }
 
     public void deleteAnimal(Long id){
         Animal animal = animalRepository.findById(id)
-                .orElseThrow(() -> new InaccessibleEntityException());
+                .orElseThrow(() -> new NotFoundException());
+
+        if(!animal.getVisitedLocations().isEmpty()){
+            throw new IllegalStateException();
+        }
 
         animalRepository.deleteById(id);
     }
